update: 1. 修改User对象实体类的字段类型 2. 同步修改init.sql中的user表字段 todo: @NotEmpty 用在集合类上面 @NotBlank 用在String上面 @NotNull    用在基本类型上

+++ src/main/java/com/suhuamo/init/enums/RoleTypeEnum.java
@@ -8,23 +8,23 @@ package com.suhuamo.init.enums;
  */
 public enum RoleTypeEnum {
     // admin-管理员；user-普通用户
-    ADMIN(1, "admin"),
-    USER(2, "user");
+    ADMIN((byte) 1, "admin"),
+    USER((byte) 2, "user");
     /**
      * 类型
      */
-    private final Integer type;
+    private final Byte type;
     /**
      * 描述
      */
     private final String desc;
 
-    RoleTypeEnum(Integer type, String desc) {
+    RoleTypeEnum(Byte type, String desc) {
         this.type = type;
         this.desc = desc;
     }
 
-    public Integer getType() {
+    public Byte getType() {
         return type;
     }
